0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,5 +1,6 @@
 #include "lists.h"
 
+
 /**
  * print_list - print counter all elements of `list_t` list
  * @h: head of linked list
@@ -8,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t i = 0;
+	int i = 0;
 
 	while (h != NULL)
 	{
